float => double to to unify with estimate.

+++ searchlib/src/vespa/searchlib/queryeval/executeinfo.h
@@ -12,11 +12,11 @@ namespace search::queryeval {
  */
 class ExecuteInfo {
 public:
-    ExecuteInfo() noexcept : ExecuteInfo(false, 1.0F, nullptr, true, true) { }
+    ExecuteInfo() noexcept : ExecuteInfo(false, 1.0, nullptr, true, true) { }
     bool isStrict() const noexcept { return _strict; }
     bool create_postinglist_when_non_strict() const noexcept { return _create_postinglist_when_non_strict; }
     bool use_estimate_for_fetch_postings() const noexcept { return _use_estimate_for_fetch_postings; }
-    float hitRate() const noexcept { return _hitRate; }
+    double hitRate() const noexcept { return _hitRate; }
     bool soft_doom() const noexcept { return _doom && _doom->soft_doom(); }
     const vespalib::Doom * getDoom() const { return _doom; }
     static const ExecuteInfo TRUE;
@@ -24,26 +24,26 @@ public:
     static ExecuteInfo create(bool strict, const ExecuteInfo & org) noexcept {
         return create(strict, org._hitRate, org);
     }
-    static ExecuteInfo create(bool strict, float hitRate, const ExecuteInfo & org) noexcept {
+    static ExecuteInfo create(bool strict, double hitRate, const ExecuteInfo & org) noexcept {
         return {strict, hitRate, org.getDoom(), org.create_postinglist_when_non_strict(), org.use_estimate_for_fetch_postings()};
     }
 
-    static ExecuteInfo create(bool strict, float hitRate, const vespalib::Doom * doom, bool postinglist_when_non_strict,
+    static ExecuteInfo create(bool strict, double hitRate, const vespalib::Doom * doom, bool postinglist_when_non_strict,
                               bool use_estimate_for_fetch_postings) noexcept
     {
          return {strict, hitRate, doom, postinglist_when_non_strict, use_estimate_for_fetch_postings};
     }
     static ExecuteInfo createForTest(bool strict) noexcept {
-        return createForTest(strict, 1.0F);
+        return createForTest(strict, 1.0);
     }
-    static ExecuteInfo createForTest(bool strict, float hitRate) noexcept {
+    static ExecuteInfo createForTest(bool strict, double hitRate) noexcept {
         return createForTest(strict, hitRate, nullptr);
     }
-    static ExecuteInfo createForTest(bool strict, float hitRate, const vespalib::Doom * doom) noexcept {
+    static ExecuteInfo createForTest(bool strict, double hitRate, const vespalib::Doom * doom) noexcept {
         return create(strict, hitRate, doom, true, true);
     }
 private:
-    ExecuteInfo(bool strict, float hitRate_in, const vespalib::Doom * doom, bool postinglist_when_non_strict,
+    ExecuteInfo(bool strict, double hitRate_in, const vespalib::Doom * doom, bool postinglist_when_non_strict,
                 bool use_estimate_for_fetch_postings) noexcept
         : _doom(doom),
           _hitRate(hitRate_in),
@@ -52,7 +52,7 @@ private:
           _use_estimate_for_fetch_postings(use_estimate_for_fetch_postings)
     { }
     const vespalib::Doom * _doom;
-    float                  _hitRate;
+    double                 _hitRate;
     bool                   _strict;
     bool                   _create_postinglist_when_non_strict;
     bool                   _use_estimate_for_fetch_postings;
